refactor(hledger): Add custom comodity with unit price
Allow ledger entries to be added with custom commodities with a unit price

Also fix an issue where key which is not available was coming with nil
values. It should not be part of map in that case. Use select-keys to facilitate
the behaviour

Update existing test cases and add new ones