fix(sms): Dynamic function dispatch
We are getting error for formatting date. Interestingly calling the correct
function on a message will parse it without issues.
The logic to select the function is also correct but when we call the function
then we get the error.
Didn't find the root cause so move forward switched to multimethods

Updated the parser logic for extraction of date time values. For pluxee they
send single digits